Move docker testing to prevent cyclic imports.
Having docker tests fixtures in opengever.testing caused cyclic imports
in policy tests:

Traceback (most recent call last):
  File "/dev/opengever.ftw/opengever/ftw/tests/test_harvest_timer.py", line 5, in <module>
    from opengever.ftw.testing import OPENGEVER_FTW_FUNCTIONAL_TESTING
  File "/dev/opengever.ftw/opengever/ftw/testing.py", line 3, in <module>
    from opengever.core.testing import functional_session_factory
  File "/dev/opengever.ftw/src/opengever.core/opengever/core/testing.py", line 28, in <module>
    from opengever.testing.docker import MSGCONVERT_SERVICE_FIXTURE
  File "/dev/opengever.ftw/src/opengever.core/opengever/testing/__init__.py", line 10, in <module>
    from opengever.core.testing import MEMORY_DB_LAYER
ImportError: cannot import name MEMORY_DB_LAYER

Moving it into opengever.core prevents the cycles.
